typecast the argument to isspace() to an int to prevent warnings on some compilers

+++ lib/transfer.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: transfer.c,v 1.140 2003-01-30 05:04:02 bagder Exp $
+ * $Id: transfer.c,v 1.141 2003-01-30 06:06:24 bagder Exp $
  ***************************************************************************/
 
 #include "setup.h"
@@ -573,7 +573,7 @@ CURLcode Curl_readwrite(struct connectdata *conn,
 
               if(end) {
                 /* skip all trailing space letters */
-                for(; isspace(*end) && (end > start); end--);
+                for(; isspace((int)*end) && (end > start); end--);
 
                 /* get length of the type */
                 len = end-start+1;
